requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
APScheduler>=3.10.0
beautifulsoup4>=4.12.0
aiohttp-socks>=0.8.0
openai>=1.0.0
